Pull out JS tooling
This tooling proved to be more difficulty than its worth.

Since vegphilly is updated intermittently, nodejs dependencies are a bad
fit because they break and/or become outdated so easily. Also, the JS is
small enough that old-fashioned module abstractions are
satisfactory. It's easier to face some frustration at the source level
than have an additional layer of tooling for features we don't
necessarily need.